# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: TopAgents API
  description: |
    ## TopAgents.ca — Real Estate Agent Review & Referral Platform

    ### Authentication
    Protected endpoints require a `Bearer` JWT token in the `Authorization` header.
    Get tokens via `POST /api/v1/auth/login`. Refresh via `POST /api/v1/auth/refresh`.

    ### Rate Limiting
    - General: 100 req / 15 min
    - Auth endpoints: 10 req / 15 min
    - Search: 60 req / 1 min

    ### Pagination
    All list endpoints accept `?page=1&limit=20`.
    Response includes `meta: { total, page, limit, totalPages, hasNext, hasPrev }`.

    ### Response Envelope
    ```json
    { "success": true, "data": { ... } }
    ```
    ```json
    { "success": false, "error": { "code": 422, "message": "Validation failed", "details": [] } }
    ```
  version: 1.0.0
  contact:
    name: TopAgents Dev Team
    email: dev@topagents.ca
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.topagents.ca/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Review:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: integer
        platform:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          nullable: true
        text:
          type: string
          nullable: true
        reviewer_name:
          type: string
        verified:
          type: boolean
        transaction_verified:
          type: boolean
        review_date:
          type: string
          format: date
        agent_responded:
          type: boolean
        agent_response:
          type: string
          nullable: true
        response_date:
          type: string
          format: date-time
          nullable: true
        property_address:
          type: string
          nullable: true
        transaction_type:
          type: string
          nullable: true
        rating_communication:
          type: integer
          nullable: true
        rating_negotiation:
          type: integer
          nullable: true
        rating_market_knowledge:
          type: integer
          nullable: true
        rating_responsiveness:
          type: integer
          nullable: true
        rating_pricing:
          type: integer
          nullable: true
        rating_paperwork:
          type: integer
          nullable: true
        rating_availability:
          type: integer
          nullable: true
        rating_recommendation:
          type: integer
          nullable: true
        imported_at:
          type: string
          format: date-time
    ReviewCreate:
      type: object
      required: [agent_id, rating, reviewer_name, reviewer_email]
      properties:
        agent_id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewer_name:
          type: string
          maxLength: 100
        reviewer_email:
          type: string
          format: email
        title:
          type: string
          maxLength: 200
        text:
          type: string
          maxLength: 5000
        platform:
          type: string
          enum: [native, google, rew, ratemyagent, ratemyagentcom, facebook, zillow]
          default: native
        platform_review_id:
          type: string
        platform_url:
          type: string
          format: uri
        property_address:
          type: string
        transaction_type:
          type: string
          enum: [buy, sell, rent]
        review_date:
          type: string
          format: date
        rating_communication:
          type: integer
          minimum: 1
          maximum: 5
        rating_negotiation:
          type: integer
          minimum: 1
          maximum: 5
        rating_market_knowledge:
          type: integer
          minimum: 1
          maximum: 5
        rating_responsiveness:
          type: integer
          minimum: 1
          maximum: 5
        rating_pricing:
          type: integer
          minimum: 1
          maximum: 5
        rating_paperwork:
          type: integer
          minimum: 1
          maximum: 5
        rating_availability:
          type: integer
          minimum: 1
          maximum: 5
        rating_recommendation:
          type: integer
          minimum: 1
          maximum: 5
...
  responses:
...
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /agents/{agentId}/reviews:
    get:
      tags: [Agents]
      summary: Get paginated reviews for an agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
        - in: query
          name: rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - in: query
          name: platform
          schema:
            type: string
        - in: query
          name: verified
          schema:
            type: boolean
        - in: query
          name: sort
          schema:
            type: string
            enum: [recent, highest, lowest]
            default: recent
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit a new review for an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Review created (pending moderation for native reviews)
        '404':
          description: Agent not found
        '409':
          description: Duplicate review
        '422':
          $ref: '#/components/responses/ValidationError'
  /reviews/import:
    post:
      tags: [Reviews]
      summary: Bulk import reviews from an external platform
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, reviews]
              properties:
                platform:
                  type: string
                  enum: [google, rew, ratemyagent, ratemyagentcom, facebook, zillow]
                reviews:
                  type: array
                  items:
                    $ref: '#/components/schemas/ReviewCreate'
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      imported:
                        type: integer
                      skipped:
                        type: integer
                      total:
                        type: integer
  /reviews/{reviewId}:
    get:
      tags: [Reviews]
      summary: Get a single review by ID
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Review object
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
  /reviews/{reviewId}/reply:
    post:
      tags: [Reviews]
      summary: Agent replies to a review (awards 10 points)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: string
                  maxLength: 500
                  example: Thank you for your kind words!
      responses:
        '200':
          description: Reply saved and 10 points awarded
        '403':
          description: You do not own this agent profile
        '404':
          $ref: '#/components/responses/NotFound'
  /reviews/{reviewId}/report:
    post:
      tags: [Reviews]
      summary: Report a review for moderation
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  enum: [fake, inappropriate, wrong_person, spam, other]
                description:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Review reported for moderation
  /reviews/{reviewId}/verify:
    patch:
      tags: [Reviews]
      summary: Admin verifies a native review and awards 50 points to the agent
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Review verified and points awarded
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── REFERRALS ─────────────────────────────────────────────────────────────
...